+++ app/secure_store.tsx
@@ -1,95 +1,32 @@
-import React, { useState } from "react";
+import React from "react";
 import {
   Text,
   View,
-  TextInput,
-  StyleSheet,
-  TouchableOpacity,
-  Alert,
   KeyboardAvoidingView,
   Platform,
   ScrollView
 } from "react-native";
-import { saveValue, getValue, deleteValue } from "../util/secure_store";
+import { useSecureStorage } from "../hooks/useSecureStorage";
+import { SecureStorageForm } from "../components/secure-storage/SecureStorageForm";
+import { SecureStorageResult } from "../components/secure-storage/SecureStorageResult";
+import { styles } from "../components/secure-storage/styles";
 
 /**
  * Demo screen component for secure storage operations.
  * Allows saving, retrieving, and deleting values from secure storage.
  */
 export default function SecureStorageScreen() {
-  const [key, setKey] = useState("");
-  const [value, setValue] = useState("");
-  const [storedValue, setStoredValue] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  /**
-   * Handles saving a key-value pair to secure storage.
-   * Validates that both key and value are provided before saving.
-   */
-  const handleSave = async () => {
-    if (!key.trim() || !value.trim()) {
-      Alert.alert("Error", "Both key and value are required");
-      return;
-    }
-
-    try {
-      setIsLoading(true);
-      await saveValue(key, value);
-      Alert.alert("Success", "Value saved securely");
-      setValue("");
-    } catch (error) {
-      Alert.alert("Error", "Failed to save value");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  /**
-   * Retrieves a value from secure storage using the provided key.
-   * Updates the storedValue state with the result.
-   */
-  const handleRetrieve = async () => {
-    if (!key.trim()) {
-      Alert.alert("Error", "Key is required");
-      return;
-    }
-
-    try {
-      setIsLoading(true);
-      const result = await getValue(key);
-      setStoredValue(result);
-
-      if (result === null) {
-        Alert.alert("Info", "No value found for this key");
-      }
-    } catch (error) {
-      Alert.alert("Error", "Failed to retrieve value");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  /**
-   * Deletes a value from secure storage using the provided key.
-   * Resets the storedValue state on successful deletion.
-   */
-  const handleDelete = async () => {
-    if (!key.trim()) {
-      Alert.alert("Error", "Key is required");
-      return;
-    }
-
-    try {
-      setIsLoading(true);
-      await deleteValue(key);
-      setStoredValue(null);
-      Alert.alert("Success", "Value deleted");
-    } catch (error) {
-      Alert.alert("Error", "Failed to delete value");
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const {
+    key,
+    setKey,
+    value,
+    setValue,
+    storedValue,
+    isLoading,
+    handleSave,
+    handleRetrieve,
+    handleDelete
+  } = useSecureStorage();
 
   return (
     <KeyboardAvoidingView
@@ -100,129 +37,20 @@ export default function SecureStorageScreen() {
         <View style={styles.content}>
           <Text style={styles.title}>Secure Storage Demo</Text>
 
-          <View style={styles.inputContainer}>
-            <Text style={styles.label}>Key:</Text>
-            <TextInput
-              style={styles.input}
-              value={key}
-              onChangeText={setKey}
-              placeholder="Enter storage key"
-            />
-          </View>
+          <SecureStorageForm
+            key={key}
+            value={value}
+            onKeyChange={setKey}
+            onValueChange={setValue}
+            onSave={handleSave}
+            onRetrieve={handleRetrieve}
+            onDelete={handleDelete}
+            isLoading={isLoading}
+          />
 
-          <View style={styles.inputContainer}>
-            <Text style={styles.label}>Value:</Text>
-            <TextInput
-              style={styles.input}
-              value={value}
-              onChangeText={setValue}
-              placeholder="Enter value to store"
-            />
-          </View>
-
-          <View style={styles.buttonContainer}>
-            <TouchableOpacity
-              style={styles.button}
-              onPress={handleSave}
-              disabled={isLoading}
-            >
-              <Text style={styles.buttonText}>Save</Text>
-            </TouchableOpacity>
-
-            <TouchableOpacity
-              style={styles.button}
-              onPress={handleRetrieve}
-              disabled={isLoading}
-            >
-              <Text style={styles.buttonText}>Retrieve</Text>
-            </TouchableOpacity>
-
-            <TouchableOpacity
-              style={styles.button}
-              onPress={handleDelete}
-              disabled={isLoading}
-            >
-              <Text style={styles.buttonText}>Delete</Text>
-            </TouchableOpacity>
-          </View>
-
-          {storedValue !== null && (
-            <View style={styles.resultContainer}>
-              <Text style={styles.resultLabel}>Retrieved Value:</Text>
-              <Text style={styles.resultValue}>{storedValue}</Text>
-            </View>
-          )}
+          <SecureStorageResult storedValue={storedValue} />
         </View>
       </ScrollView>
     </KeyboardAvoidingView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  scrollContent: {
-    flexGrow: 1,
-  },
-  content: {
-    flex: 1,
-    padding: 20,
-    justifyContent: "center",
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: "bold",
-    marginBottom: 30,
-    textAlign: "center",
-  },
-  inputContainer: {
-    marginBottom: 16,
-  },
-  label: {
-    fontSize: 16,
-    marginBottom: 8,
-    fontWeight: "500",
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: "#ccc",
-    borderRadius: 8,
-    padding: 12,
-    fontSize: 16,
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginTop: 20,
-    marginBottom: 30,
-  },
-  button: {
-    backgroundColor: "#007AFF",
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    borderRadius: 8,
-    flex: 1,
-    marginHorizontal: 4,
-    alignItems: "center",
-  },
-  buttonText: {
-    color: "white",
-    fontWeight: "bold",
-    fontSize: 16,
-  },
-  resultContainer: {
-    backgroundColor: "#f5f5f5",
-    padding: 16,
-    borderRadius: 8,
-    marginTop: 20,
-  },
-  resultLabel: {
-    fontSize: 16,
-    fontWeight: "500",
-    marginBottom: 8,
-  },
-  resultValue: {
-    fontSize: 16,
-  },
-});
